Rearrange Game19 bit addresses.
- Give ID more bits (44 instead of 32).
- Change 'int' IDs to longs in several places.
- Clarify role of flags, type, and ID components.
- Add a BitAddresses.toString function.
- Convert line endings in Message and BitAddresses sources,
  which had extensive modifications anyway.

+++ src/togos/networkrts/experimental/game19/world/IDGenerator.java
@@ -1,15 +1,15 @@
 package togos.networkrts.experimental.game19.world;
 
 public class IDGenerator {
-	int next;
+	long next;
 	
-	public IDGenerator( int startAt ) {
+	public IDGenerator( long startAt ) {
 		next = startAt;
 	}
 	
 	public IDGenerator() { this(1); }
 	
-	public synchronized int newId() {
+	public synchronized long newId() {
 		return next++;
 	}
 }
